Move application insights tracking to shared infrastructure
