--- application/account-management/WebApp/shared/lib/applicationInsights/ErrorFallback.tsx
@@ -1,9 +1,0 @@
-import { Trans } from "@lingui/macro";
-
-export function ErrorFallback() {
-  return (
-    <h1>
-      <Trans>Something went wrong, please try again</Trans>
-    </h1>
-  );
-}
