+++ application/account-management/WebApp/shared/translations/locale/en-US.po
@@ -107,9 +107,6 @@ msgstr "Sign in with your company email address to get started building on Platf
 msgid "Sign up"
 msgstr "Sign up"
 
-msgid "Something went wrong, please try again"
-msgstr "Something went wrong, please try again"
-
 msgid "Subdomain"
 msgstr "Subdomain"
