+++ application/shared-webapp/infrastructure/package.json
@@ -18,6 +18,7 @@
   },
   "devDependencies": {
     "@repo/config": "*",
+    "@repo/build": "*",
     "@repo/utils": "*"
   }
 }
